Extract orphan file removal logic
